Separate packaging tests into new file
